refactor(api): move key utilities to utils/crypto.ts
