+++ server/api/test/utils/crypto.ts
@@ -8,6 +8,50 @@ import {
   signatureKeySaltLength,
 } from '../../src/constants'
 
+export type SerializedSignatureKeys = {
+  publicKey: string
+  privateKey: string
+}
+
+export type SerializedKeys = {
+  encryptionKey: string
+  signatureKeys: SerializedSignatureKeys
+  iv: string
+  salt: string
+}
+
+const isSerializedSignatureKeys = (
+  obj: unknown
+): obj is SerializedSignatureKeys => {
+  if (typeof obj !== 'object' || obj === null) {
+    return false
+  }
+
+  return (
+    'publicKey' in obj &&
+    typeof obj.publicKey === 'string' &&
+    'privateKey' in obj &&
+    typeof obj.privateKey === 'string'
+  )
+}
+
+const isSerializedKeys = (obj: unknown): obj is SerializedKeys => {
+  if (typeof obj !== 'object' || obj === null) {
+    return false
+  }
+
+  return (
+    'encryptionKey' in obj &&
+    typeof obj.encryptionKey === 'string' &&
+    'signatureKeys' in obj &&
+    isSerializedSignatureKeys(obj.signatureKeys) &&
+    'iv' in obj &&
+    typeof obj.iv === 'string' &&
+    'salt' in obj &&
+    typeof obj.salt === 'string'
+  )
+}
+
 export const importKey = async (password: string) => {
   const encoder = new TextEncoder()
 
@@ -63,4 +107,38 @@ export const getSignature = async (
   )
 
   return signature
+}
+
+export const decryptSerializedKeys = async (
+  encryptedKeys: string,
+  passkeySecret: string,
+  ivString: string,
+  saltString: string
+) => {
+  const iv = Buffer.from(ivString, 'base64')
+  const salt = Buffer.from(saltString, 'base64')
+  const decoder = new TextDecoder()
+
+  const importedKey = await importKey(passkeySecret)
+  const derivedKey = await deriveKey(importedKey, salt)
+
+  const encryptedKeysBuffer = Buffer.from(encryptedKeys, 'base64')
+
+  const decryptedKeysBuffer = await crypto.subtle.decrypt(
+    {
+      name: contentEncryptionAlgorithmName,
+      iv,
+    },
+    derivedKey,
+    encryptedKeysBuffer
+  )
+
+  const decryptedKeysString = decoder.decode(decryptedKeysBuffer)
+  const decryptedKeys = JSON.parse(decryptedKeysString)
+
+  if (!isSerializedKeys(decryptedKeys)) {
+    throw new Error()
+  }
+
+  return decryptedKeys
 }
